Preparations to implement local import buttons
